binai/requirements.txt
# Binance API
python-binance

# Veri işleme ve teknik analiz (Gelecekteki stratejiler için eklendi)
numpy
pandas


# HTTP talepleri
requests

# Loglama için
coloredlogs